fixed cast time being in ticks instead of seconds

+++ src/main/java/de/raidcraft/skills/api/persistance/SkillProperties.java
@@ -55,7 +55,7 @@ public interface SkillProperties {
 
     public double getDamageLevelModifier();
 
-    public int getCastTime();
+    public double getCastTime();
 
     public double getCastTimeLevelModifier();
